Service1/requirements.txt
SQLAlchemy==1.3.13
pykafka==2.8.0
Flask_Cors==3.0.8
connexion==2.6.0
requests==2.21.0
PyYAML==5.3
